+++ elements/user-picker/src/components/smart-user-picker/config/index.ts
@@ -36,3 +36,5 @@ export const setSmartUserPickerEnv = (newEnv: 'prod' | 'local') =>
   (env = newEnv);
 
 export const getConfig = () => (env === 'local' ? LOCAL_CONFIG : PRD_CONFIG);
+
+export const JDOG_CLOUD_ID = '49d8b9d6-ee7d-4931-a0ca-7fcae7d1c3b5';