+++ design-system/inline-edit/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/inline-edit
 
+## 12.0.6
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 12.0.5
 
 ### Patch Changes